Basic refactoring
