+++ udmis/src/main/java/com/google/bos/udmi/service/messaging/MessageBase.java
@@ -1,32 +1,19 @@
 package com.google.bos.udmi.service.messaging;
 
-import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.udmi.util.JsonUtil.convertTo;
 import static com.google.udmi.util.JsonUtil.fromString;
 
 import com.google.bos.udmi.service.pod.ContainerBase;
-import com.google.common.collect.BiMap;
-import com.google.common.collect.HashBiMap;
-import com.google.common.collect.ImmutableMap;
 import com.google.udmi.util.Common;
-import java.util.AbstractMap.SimpleEntry;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.LinkedBlockingDeque;
-import org.jetbrains.annotations.NotNull;
+import java.util.function.Consumer;
 import udmi.schema.Envelope;
-import udmi.schema.Envelope.SubFolder;
-import udmi.schema.Envelope.SubType;
-import udmi.schema.SystemState;
 
 /**
  * Base class for supporting a variety of messaging interfaces.
@@ -35,105 +22,17 @@ public abstract class MessageBase extends ContainerBase implements MessagePipe {
 
   private static final String DEFAULT_NAMESPACE = "default-namespace";
   private static final String LOOP_EXIT_MARK = "loop-exit";
-  private static final String DEFAULT_HANDLER = "event/null";
-  private static final String EXCEPTION_HANDLER = "exception_handler";
-  private static final Map<SimpleEntry<SubType, SubFolder>, Class<?>> SPECIAL_TYPES =
-      ImmutableMap.of(getTypeFolderEntry(SubType.STATE, SubFolder.UPDATE), StateUpdate.class);
-  private static final Map<Class<?>, SimpleEntry<SubType, SubFolder>> CLASS_TYPES = new HashMap<>();
-  private static final BiMap<String, Class<?>> TYPE_CLASSES = HashBiMap.create();
-
-  static {
-    initializeHandlerTypes();
-  }
 
   private final ExecutorService executor = Executors.newSingleThreadExecutor();
-  private final Map<String, MessageHandler<Object>> handlers = new HashMap<>();
-  private final Map<Object, Envelope> messageEnvelopes = new ConcurrentHashMap<>();
-  private final Map<Class<?>, String> specialClasses = ImmutableMap.of(
-      Object.class, DEFAULT_HANDLER,
-      Exception.class, EXCEPTION_HANDLER
-  );
   BlockingQueue<String> sourceQueue;
   private Future<Void> sourceFuture;
-
-  private static String getMapKey(SubType subType, SubFolder subFolder) {
-    SubType useType = Optional.ofNullable(subType).orElse(SubType.EVENT);
-    return String.format("%s/%s", useType, subFolder);
-  }
-
-  private static Class<?> getMessageClass(SubType type, SubFolder folder) {
-    String typeName = Common.capitalize(folder.value()) + Common.capitalize(type.value());
-    String className = SystemState.class.getPackageName() + "." + typeName;
-    try {
-      return Class.forName(className);
-    } catch (ClassNotFoundException e) {
-      return null;
-    }
-  }
-
-  @NotNull
-  private static SimpleEntry<SubType, SubFolder> getTypeFolderEntry(SubType type,
-      SubFolder folder) {
-    return new SimpleEntry<>(type, folder);
-  }
-
-  private static void initializeHandlerTypes() {
-    Arrays.stream(SubType.values()).forEach(type -> Arrays.stream(SubFolder.values())
-        .forEach(folder -> registerHandlerType(type, folder)));
-    SPECIAL_TYPES.forEach(
-        (entry, clazz) -> registerMessageClass(entry.getKey(), entry.getValue(), clazz));
-  }
-
-  /**
-   * Make a new message bundle for the given object, inferring the type and folder from the class
-   * itself (using the predefined lookup map).
-   */
-  public static Bundle makeMessageBundle(Object message) {
-    if (message instanceof Bundle || message == null) {
-      return (Bundle) message;
-    }
-    Bundle bundle = new Bundle();
-    bundle.message = message;
-    bundle.envelope = new Envelope();
-    SimpleEntry<SubType, SubFolder> messageType = CLASS_TYPES.get(message.getClass());
-    checkNotNull(messageType, "type entry not found for " + message.getClass());
-    bundle.envelope.subType = messageType.getKey();
-    bundle.envelope.subFolder = messageType.getValue();
-    return bundle;
-  }
+  private Consumer<Bundle> dispatcher;
 
   static String normalizeNamespace(String configSpace) {
     return Optional.ofNullable(configSpace).orElse(DEFAULT_NAMESPACE);
   }
 
-  private static void registerHandlerType(SubType type, SubFolder folder) {
-    Class<?> messageClass = getMessageClass(type, folder);
-    registerMessageClass(type, folder, messageClass);
-  }
-
-  private static void registerMessageClass(SubType type, SubFolder folder, Class<?> messageClass) {
-    String mapKey = getMapKey(type, folder);
-    if (messageClass != null) {
-      TYPE_CLASSES.put(mapKey, messageClass);
-      CLASS_TYPES.put(messageClass, getTypeFolderEntry(type, folder));
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  protected Future<Void> handleQueue(BlockingQueue<String> queue) {
-    return (Future<Void>) executor.submit(() -> this.messageLoop(queue));
-  }
-
-  protected abstract void publishBundle(Bundle messageBundle);
-
-  private void ignoreMessage(Object message) {
-  }
-
-  private Envelope makeExceptionEnvelope() {
-    return new Envelope();
-  }
-
-  private void messageLoop(BlockingQueue<String> queue) {
+  void messageLoop(BlockingQueue<String> queue) {
     try {
       while (true) {
         try {
@@ -144,9 +43,9 @@ public abstract class MessageBase extends ContainerBase implements MessagePipe {
             return;
           }
           Bundle bundle = extractBundle(bundleString);
-          processMessage(bundle);
+          dispatcher.accept(bundle);
         } catch (Exception e) {
-          processMessage(makeExceptionEnvelope(), EXCEPTION_HANDLER, e);
+          dispatcher.accept(makeExceptionBundle(e));
         }
       }
     } catch (Exception loopException) {
@@ -154,33 +53,23 @@ public abstract class MessageBase extends ContainerBase implements MessagePipe {
     }
   }
 
-  private void processMessage(Envelope envelope, String mapKey, Object messageObject) {
-    try {
-      messageEnvelopes.put(messageObject, envelope);
-      handlers.get(mapKey).accept(messageObject);
-    } finally {
-      messageEnvelopes.remove(messageObject);
-    }
+  private Bundle makeExceptionBundle(Exception e) {
+    Bundle bundle = new Bundle();
+    bundle.envelope = new Envelope();
+    bundle.message = e;
+    return bundle;
   }
 
-  void processMessage(Bundle bundle) {
-    Envelope envelope = checkNotNull(bundle.envelope, "bundle envelope is null");
-    String mapKey = getMapKey(envelope.subType, envelope.subFolder);
-    try {
-      handlers.computeIfAbsent(mapKey, key -> {
-        info("Defaulting messages of type/folder " + mapKey);
-        return handlers.getOrDefault(DEFAULT_HANDLER, this::ignoreMessage);
-      });
-      Class<?> handlerType = TYPE_CLASSES.getOrDefault(mapKey, Object.class);
-      Object messageObject = convertTo(handlerType, bundle.message);
-      processMessage(envelope, mapKey, messageObject);
-    } catch (Exception e) {
-      throw new RuntimeException("While processing message key " + mapKey, e);
-    }
+  @SuppressWarnings("unchecked")
+  protected Future<Void> handleQueue(BlockingQueue<String> queue) {
+    return (Future<Void>) executor.submit(() -> messageLoop(queue));
   }
+
+  public abstract void publishBundle(Bundle bundle);
 
   @Override
-  public void activate() {
+  public void activate(Consumer<Bundle> messageDispatcher) {
+    dispatcher = messageDispatcher;
     checkState(sourceFuture == null, "pipe already activated");
     if (sourceQueue == null) {
       sourceQueue = new LinkedBlockingDeque<>();
@@ -197,20 +86,6 @@ public abstract class MessageBase extends ContainerBase implements MessagePipe {
   @Override
   public boolean isActive() {
     return sourceFuture != null && !sourceFuture.isDone();
-  }
-
-  @Override
-  public void publish(Object message) {
-    publishBundle(makeMessageBundle(message));
-  }
-
-  @Override
-  @SuppressWarnings("unchecked")
-  public <T> void registerHandler(Class<T> clazz, MessageHandler<T> handler) {
-    String mapKey = specialClasses.getOrDefault(clazz, TYPE_CLASSES.inverse().get(clazz));
-    if (handlers.put(mapKey, (MessageHandler<Object>) handler) != null) {
-      throw new RuntimeException("Type handler already defined for " + mapKey);
-    }
   }
 
   /**
