+++ udmis/src/main/java/com/google/bos/udmi/service/messaging/MessagePipe.java
@@ -2,9 +2,8 @@ package com.google.bos.udmi.service.messaging;
 
 import static com.google.common.base.Preconditions.checkState;
 
+import com.google.bos.udmi.service.messaging.MessageBase.Bundle;
 import com.google.common.collect.ImmutableMap;
-import java.util.AbstractMap.SimpleEntry;
-import java.util.Collection;
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -33,56 +32,14 @@ public interface MessagePipe {
   }
 
   /**
-   * Static factory method for creating handler specifications.
-   */
-  static <T> HandlerSpecification messageHandlerFor(Class<T> clazz, MessageHandler<T> consumer) {
-    return new HandlerSpecification(clazz, consumer);
-  }
-
-  /**
    * Activate the receive loop of the message handler. Usually after handlers are registered!
    */
-  void activate();
+  void activate(Consumer<Bundle> callback);
 
   /**
    * Check if this pipe has been activated (and is still active).
    */
   boolean isActive();
 
-  /**
-   * Publish a message to the outgoing channel of this pipe.
-   */
-  void publish(Object message);
-
-  <T> void registerHandler(Class<T> targetClass, MessageHandler<T> handler);
-
-  /**
-   * Convenience function to register an entire collection of handler specifications.
-   */
-  default void registerHandlers(Collection<HandlerSpecification> messageHandlers) {
-    messageHandlers.forEach(handler -> handler.registerWith(this));
-  }
-
-  /**
-   * Represent a type-happy consumer into a more generic functional specification.
-   */
-  interface MessageHandler<T> extends Consumer<T> {
-
-  }
-
-  /**
-   * Represent a type-happy consumer into a more generic specification. No actual logic, just makes
-   * calling code cleaner and less cluttered with java-type crazyness.
-   */
-  class HandlerSpecification extends SimpleEntry<Class<?>, MessageHandler<?>> {
-
-    public <T> HandlerSpecification(Class<T> clazz, MessageHandler<T> handler) {
-      super(clazz, handler);
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> void registerWith(MessagePipe pipe) {
-      pipe.registerHandler((Class<T>) getKey(), (MessageHandler<T>) getValue());
-    }
-  }
+  void publishBundle(Bundle bundle);
 }
